requirements.txt
roboflow==1.1.31
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
matplotlib==3.8.3
streamlit==1.31.1
torch==2.2.0
torchvision==0.17.0
supervision==0.19.0
requests==2.31.0